Introduce service layer for business orchestration
Create SentryService that owns all sentry business logic:
- Coordinates SentryRepository (data access) and IEventPublisher
- Publishes all domain events
- Exposes public methods: StartSentryAsync, FinishSentryAsync

Refactor SentryRepository to pure data access:
- InsertSentryAsync, GetSentryAsync, UpdateSentryEndAsync, GetActiveSentry
- No event publishing — repository owns only persistence
- Simplified and testable in isolation

Update GraphQL mutation to call SentryService instead of repository:
- Mutation remains thin adapter: validate input → call service → return result
- Business logic moves out of GraphQL layer

Benefits:
- Business logic is reusable (could be called from REST, gRPC, CLI)
- Easier to test business logic without GraphQL context
- Clear separation: GraphQL (presentation) → Service (business) → Repository (data)
- Ready for mid-shift mutations (addGuard, removeGuard, etc.)
- Pattern for future modules to follow

Register SentryService in DI as singleton, alongside repositories.
